Add resources
We add our whole infrastructure in Terraform configuration. Our ultimate goal is
to serve wordpress from our vm instance. For that we also need an RDS db, an elastic load balancer and two security groups, one for the vm instance and one for the lb in order to accept traffic from the Internet.