Massive update main.tf
This is basically our first complete attempt to create and configure our
whole infrastructure. We add the resources we were lacking to create the
proper connection between our lb and vm. We create the whole procedure
to get a TLS certificate for our domain. We also generate a random
password for our db instance for security reasons. Still a couple of
issues though with the default subnets. Still missing Route53 A records.